+++ FoxSky.Img/FileProcessors/Mode.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FoxSky.Img.FileProcessors
+{
+    public enum Mode
+    {
+        Move,
+        Copy
+    }
+}
