+++ FoxSky.Img/Utilities/Logger.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FoxSky.Img.Utilities
+{
+    public class Logger
+    {
+        public static void LogSuccess(string message)
+        {
+            try
+            {
+                Console.Write($"[{DateTime.Now}]");
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.Write("Success! ");
+                Console.Write($"{message}");
+                Debug.WriteLine(message);
+                Console.WriteLine();
+            }
+            finally
+            {
+                Console.ResetColor();
+            }
+        }
+
+        public static void LogError(string message)
+        {
+            try
+            {
+                Console.Write($"[{DateTime.Now}]");
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.Write("Error! ");
+                Console.Write($"{message}");
+                Debug.WriteLine(message);
+                Console.WriteLine();
+            }
+            finally
+            {
+                Console.ResetColor();
+            }
+        }
+    }
+}
